safe-print: add exception-guarded `sprint` helper
This is a drop-in replacement for the `print` function, added because when jobs were
started and then the controlling terminal closed, they might error out at the end of
the job (if not running with --quiet, or if there was an error message to print out
from the message queue) with an IOError.

In the case of IOError from `print`, just discard it and continue.  For all other
exceptions, re-raise.