general: update for python3 compat
Started with python-modernize, and then added a number of manual changes to get the
tests and linters passing.

Also ran a few manual tests to make sure that everything was happy. The database
format is still compatible, and a single install should work for either py2 or py3,
at this point.