ui: fix '--pid' for python3 and improve '--int' wrt sudo
--pid was not decoding the output from pstree, so it was printing out a byte-array
repr instead. Use the newly added autoDecode for the pstree output.

--int had some problems re-exec'ing python under sudo for virtualenv
installations. We can use sys.executable to ensure that we run the helper kill script
inside the same virtualenv (by reusing sys.executable) so that we're loading the
right jobrunner modules.

Updated the interrupt_running_test to check the --pid output, and also to ensure that
we can actually run the "sudo" test when sudo is available.

The problem causing failures in the integration test before was, presumably, that we
were too quick to kill the process with SIGTERM which means that it will not be able
to update the job db to remove the running job, and so the test thinks the job is
still running, even though it was actually killed. To get around this, send more
SIGINT, first, and then finally, SIGTERM, SIGKILL.