utils: fix autoDecode error for specific sequence
Since we're probably going to be seeing either ASCII or UTF-8 input anyway, bump the
detection confidence requirement from 0.5 to 0.8. In the case of the first test input
shown in utils_test.py, it was coming in at 0.559 for Windows-1254 and just 0.505 for
UTF-8, when in fact, it's UTF-8.

This also makes me question using chardet at all, but it probably won't hurt at the
new confidence threshold.