info: only resolve workspaceIdentity, workspaceProject as needed
The plugins' implementation of these methods could be somewhat expensive, so we don't
want to, for example, call workspaceProject() each time a JobInfo is initialized as
part of reading a job entry from the database.

To address this, we call resolve() to resolve the attributes' values in only two
instances:
  1. when the job is created (but not yet started), and
  2. when the job is started.

For the case of (2), `resolve(force=True)` is used so that we override the creation
values in case something has changed in the environment that might affect these values.