main: used lockedSection instead of explicit lock/unlock calls
It's better to use a context manager for locks so that unlock calls are not missed.
Basically we just put a locked section around the implementation of "main", but back
out of it temporarily when we (maybe) fork the child process that runs the job. The
parent then exits (lock not held), while the child process will regain the lock while
it handles the rest of the options, finally releasing it before calling runJob.

In runJob, again, when it terminates, it updates the database and so uses another
lockedSection() call.