plugins: allow plugins to provide a relative priority value
The highest priority for a plugin is 0, the lowest is 1<<31. If there are multiple
plugins providing a function at the same level, then it's down to the implementation
in Plugins itself. Currently only getResources() combines the output from multiple
plugins. workspaceIdentity() and workspaceProject() both return the response from the
first plugin at a given level.