db: add debug logging of queries
Since there's still a lot of overhead checking if DEBUG logging is enabled when many
calls are involved, we check once and cache the value in the SqliteJobRepository
initializer.

I had to move up the "debug logger" creation function in main.py so that this check
happens after logging has been set up.